+++ src/components/molecules/Select/Components/Modal/index.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import {create} from 'react-test-renderer';
-import Text from 'atoms/Text';
+import Typography from 'atoms/Typography';
 import BaseButton from 'atoms/BaseButton';
 import Modal from './';
 
 const validData = {
 	show: true,
 	setShow: () => {},
-	children: <Text>Option 1</Text>,
+	children: <Typography>Option 1</Typography>,
 	measures: {
 		width: 0,
 		pageY: 0,